Move ModelWithVars into separate module
